+++ code/Model/Observer.php
@@ -179,4 +179,20 @@ class Algolia_Algoliasearch_Model_Observer
         Mage::getResourceModel('algoliasearch/fulltext')->rebuildIndex($storeId, $productIds);
         return $this;
     }
+
+    public function prepareLayoutBefore(Varien_Event_Observer $observer)
+    {
+        /* @var $block Mage_Page_Block_Html_Head */
+        $block = $observer->getEvent()->getBlock();
+
+        if ("head" == $block->getNameInLayout()) {
+            $block->addJs('../skin/frontend/base/default/algoliasearch/jquery.min.js');
+            $block->addJs('../skin/frontend/base/default/algoliasearch/jquery-ui.js');
+            $block->addJs('../skin/frontend/base/default/algoliasearch/typeahead.min.js');
+            $block->addJs('../skin/frontend/base/default/algoliasearch/jquery.noconflict.js');
+            $block->addCss('algoliasearch/jquery-ui.min.css');
+        }
+
+        return $this;
+    }
 }